fix(segment): content-sensitive segment checksum — XXH64 replaces content-blind rolling CRC32 (gastrolog-1vepg0)
The segment record checksum rolled CRC32/IEEE over lenPrefix ++ body ++
bodyCRC per frame. Each frame ends with its own CRC32, and by CRC
linearity CRC(M ++ CRC(M)) cancels the content contribution regardless
of prior rolling state — the checksum pinned only frame-length
structure (count, lengths, truncation), not content. A holder serving
corrupted-in-place or substituted bytes with matching frame geometry
and fixed-up frame CRCs passed segment.Open, publish, and collection's
published-checksum verify (gastrolog-5zotim), then merged into home
GLCBs — potentially divergent GLCBs across homes. Equal-length frame
reorder was equally invisible.

Fix: SegmentChecksum is now XXH64 (cespare/xxhash/v2, already in the
dependency tree; promoted to a direct require) over the same contiguous
record region [HeaderSize:IndexOffset). A non-linear digest has no
cancellation, keeps the single streaming verify pass (no frame-aware
walking), and still covers each frame's embedded CRC bytes. Zero stays
the "no data / no expectation" sentinel.

Format changed in place (V1-until-v1.0.0 policy, no migration):

- segment header: SegmentChecksum uint32 -> uint64; subsequent field
  offsets and HeaderSize shift by 4
- segment writer: rolling digest in AppendFrames; reconcile, BuildIndex,
  and verifyIndexedLayout recompute/verify the same digest; index and
  source-index checksums remain CRC32 (no embedded-CRC cancellation
  there)
- vaultctlfsm proto: PublishCompletedSegmentCommand.checksum and
  CompletedSegmentEntry.checksum uint32 -> uint64 (regenerated Go + TS)
- vaultctlfsm.CompletedSegmentEntry, distribution.Metadata,
  collection.AssignedSegment, PromoteVerified: checksum widened to
  uint64; PromoteVerified's known-limitation caveat removed

Reproduction-first tests (all failed on the old CRC): same-geometry
segments with different bytes shared checksum a7a5af15; a same-length
substitution with fixed-up frame CRC passed Open and PromoteVerified.
Now pinned: content sensitivity, substitution rejected at Open and at
PromoteVerified, equal-length frame reorder rejected, and a tampered
holder copy rejected on the collection pull path with convergence once
the holder serves authentic bytes.

+++ backend/internal/pipeline/collection/log.go
@@ -10,7 +10,7 @@ import (
 type AssignedSegment struct {
 	VaultID   glid.GLID
 	SegmentID glid.GLID
-	Checksum  uint32
+	Checksum  uint64
 }
 
 // LogReader rolls the vault-ctl log and returns segments assigned to this home.
